+++ src/components/self-loading-button.vue
